Working on batching

+++ src/traintestgyms/ignitegym.py
@@ -60,7 +60,7 @@ class TrainTestGym:
 
     ### Training step
     def __train_step(self, engine, batch):
-        batch = batch.to(self.device, dtype=torch.float32)
+        batch = batch.to(self.device)
 
         if engine.t_start != 0:
             engine.t_start = batch[0,self.time_column_idx].item()
